requirements.txt
undetected-chromedriver==3.5.3
python-telegram-bot==20.3
tensorflow==2.15.0
selenium-wire==5.1.0
numpy==1.26.3
python-dotenv==1.0.0
schedule==1.2.0